Crate partial_fn

A partial function of type PartialFn<A,B> is a unary function whose domain is a subset of A. In addition to being able to call a PartialFn, a method is_defined_at is provided in order to test whether the given PartialFn is defined at a particular value.

Macros

partial_fn
Construct a partial function PartialFn<A,B> from a series of one or more match statements.

Structs

PartialFn
The PartialFn type.