Macro partial_fn::partial_fn

macro_rules! partial_fn {
    ($($($pat:pat)|+ $(if $cond:expr)* => $result:expr),*) => { ... };
}

Construct a partial function PartialFn<A,B> from a series of one or more match statements.

Remarks

If the match arms provided represent a total function, then you will get an "unreachable pattern" warning. This macro automatically generates a catch-all case and in the case of a total function, this catch-all case is superfluous. It can be safely ignored.

Examples

let pf: PartialFn<i32, String> = partial_fn! {
        1 => "foo".to_string(),
        2 | 3 => "bar".to_string(),
        4...5 => "baz".to_string(),
        x if x >= 6 && x <= 7 => format!("qux{}", x)
};
assert_eq!(Some("foo".to_string()), pf.call(1));
assert_eq!(Some("bar".to_string()), pf.call(2));
assert_eq!(Some("baz".to_string()), pf.call(4));
assert_eq!(Some("qux6".to_string()), pf.call(6));